% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
A numeric scalar giving the mean.
}
\description{
Computes the mean of a \code{sparse_numeric} vector, treating
missing entries as zeros (i.e., over the full length).
}
\examples{
x <- as(c(1, 0, 2, 0), "sparse_numeric")
mean(x)

}
