% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Euclidean norm of a sparse numeric vector}
\usage{
norm(x)

\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}
}
\value{
A numeric scalar giving the norm.
}
\description{
Computes the Euclidean (L2) norm of a \code{sparse_numeric} vector:
the square root of the sum of squared elements.
}
\examples{
x <- as(c(3, 0, 4), "sparse_numeric")
norm(x)

}
