% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot overlapping non-zero entries of two sparse vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A sparse_numeric vector}

\item{y}{A sparse_numeric vector of same length}

\item{...}{Extra plot args}
}
\description{
\code{plot} method for \code{sparse_numeric} vectors. It plots the values
of two sparse vectors at their overlapping non-zero positions.
}
