% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Display a sparse_numeric object}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\description{
\code{show} method for \code{sparse_numeric} prints the length,
number of non-zero entries, and the first few non-zero positions/values.
}
