% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Crossproduct of two sparse numeric vectors}
\usage{
sparse_crossprod(x, y)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{y}{A \code{sparse_numeric} vector of the same length.}
}
\value{
A numeric scalar equal to \code{sum(x * y)}.
}
\description{
Computes the dot product of two \code{sparse_numeric} vectors.
}
\examples{
x <- as(c(1, 0, 2), "sparse_numeric")
y <- as(c(0, 5, 3), "sparse_numeric")
sparse_crossprod(x, y)

}
