% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse numeric vectors element-wise}
\usage{
sparse_mult(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{y}{A \code{sparse_numeric} vector of the same length.}

\item{e1}{First \code{sparse_numeric} vector (for operator methods).}

\item{e2}{Second \code{sparse_numeric} vector (for operator methods).}
}
\value{
A \code{sparse_numeric} vector containing the element-wise product.
}
\description{
Element-wise multiplication of two \code{sparse_numeric} vectors.
}
\examples{
x <- as(c(1, 0, 2), "sparse_numeric")
y <- as(c(0, 5, 3), "sparse_numeric")
sparse_mult(x, y)

}
