% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse numeric vector class}
\description{
Stores sparse numeric vectors.
}
\section{Slots}{

\describe{
\item{\code{value}}{Non-zero values.}

\item{\code{pos}}{Positions of non-zero values.}

\item{\code{length}}{Total length including zeros.}
}}

