% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\docType{package}
\name{sparsevectorspackage-package}
\alias{sparsevectorspackage}
\alias{sparsevectorspackage-package}
\title{Homework6: Sparse Numeric Vector Tools}
\description{
Provides an S4 class for sparse numeric vectors and basic operations
such as arithmetic, norms, and standardization without converting
to dense format.
}
\author{
\strong{Maintainer}: Amanda Roberts \email{youremail@example.com}

}
\keyword{internal}
