% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric_class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse numeric vector}
\usage{
standardize(x)

\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}
}
\value{
A standardized \code{sparse_numeric} vector with (approximately)
mean 0 and standard deviation 1 when converted to dense form.
}
\description{
Standardizes a \code{sparse_numeric} vector by subtracting the mean
(computed over all entries including zeros) and dividing by the sample
standard deviation (also over all entries).
}
\examples{
x <- as(c(1, 0, 2, 0), "sparse_numeric")
z <- standardize(x)
as(z, "numeric")

}
